+++ src/frykit/plot/artist.py
@@ -25,10 +25,20 @@ from matplotlib.transforms import Affine2D, Bbox, ScaledTranslation, offset_copy
 from numpy import ma
 from numpy.typing import ArrayLike, NDArray
 from shapely.geometry.base import BaseGeometry
+from typing_extensions import Unpack
 
 from frykit.calc import get_values_between, t_to_az
 from frykit.conf import config
 from frykit.plot.projection import PLATE_CARREE
+from frykit.plot.typing import (
+    CompassPcKwargs,
+    CompassTextKwargs,
+    FrameKwargs,
+    GeometryPathCollectionKwargs,
+    QuiverLegendPatchKwargs,
+    QuiverLegendQkKwargs,
+    TextCollectionKwargs,
+)
 from frykit.plot.utils import (
     EMPTY_PATH,
     box_path,
@@ -195,7 +205,7 @@ class GeometryPathCollection(PathCollection):
         crs: CRS | None,
         fast_transform: bool | None = None,
         skip_outside: bool | None = None,
-        **kwargs: Any,
+        **kwargs: Unpack[GeometryPathCollectionKwargs],
     ) -> None:
         self.geometries = np.asarray(geometries, dtype=np.object_)
         self.fast_transform = _resolve_fast_transform(fast_transform)
@@ -269,7 +279,7 @@ class TextCollection(Artist):
         y: ArrayLike,
         s: ArrayLike,
         skip_outside: bool | None = None,
-        **kwargs: Any,
+        **kwargs: Unpack[TextCollectionKwargs],
     ) -> None:
         super().__init__()
         self.skip_outside = _resolve_skip_outside(skip_outside)
@@ -284,12 +294,14 @@ class TextCollection(Artist):
             raise ValueError("x、y 和 s 长度必须相同")
         self.coords = np.column_stack([self.x, self.y])
 
-        kwargs = normalize_kwargs(kwargs, Text)
+        kwargs = normalize_kwargs(kwargs)  # type: ignore
+        kwargs = cast(TextCollectionKwargs, kwargs)
         kwargs.setdefault("horizontalalignment", "center")
         kwargs.setdefault("verticalalignment", "center")
         kwargs.setdefault("clip_on", True)
         self.texts = [
-            Text(xi, yi, si, **kwargs) for xi, yi, si in zip(self.x, self.y, self.s)
+            Text(xi, yi, si, **kwargs)  # type: ignore
+            for xi, yi, si in zip(self.x, self.y, self.s)
         ]
 
     def set_figure(self, fig: Figure | SubFigure) -> None:
@@ -338,8 +350,8 @@ class QuiverLegend(QuiverKey):
         loc: Literal[
             "lower left", "lower right", "upper left", "upper right"
         ] = "lower right",
-        qk_kwargs: dict[str, Any] | None = None,
-        patch_kwargs: dict[str, Any] | None = None,
+        qk_kwargs: QuiverLegendQkKwargs | None = None,
+        patch_kwargs: QuiverLegendPatchKwargs | None = None,
     ) -> None:
         self.units = units
         self.width = width
@@ -369,6 +381,7 @@ class QuiverLegend(QuiverKey):
                 )
 
         qk_kwargs = normalize_kwargs(qk_kwargs)  # type: ignore
+        qk_kwargs = cast(QuiverLegendQkKwargs, qk_kwargs)
 
         super().__init__(
             Q=Q,
@@ -385,7 +398,8 @@ class QuiverLegend(QuiverKey):
         zorder = qk_kwargs.get("zorder", 5)
         self.set_zorder(zorder)
 
-        patch_kwargs = normalize_kwargs(patch_kwargs, Rectangle)  # type: ignore
+        patch_kwargs = normalize_kwargs(patch_kwargs)  # type: ignore
+        patch_kwargs = cast(QuiverLegendPatchKwargs, patch_kwargs)
         patch_kwargs.setdefault("linewidth", 0.8)
         patch_kwargs.setdefault("edgecolor", "k")
         patch_kwargs.setdefault("facecolor", "w")
@@ -407,7 +421,7 @@ class QuiverLegend(QuiverKey):
 
         # 将 qk 调整至 patch 的中心
         self._labelsep_inches: float
-        fontsize = self.text.get_fontsize() / 72  # type: ignore
+        fontsize = cast(int, self.text.get_fontsize()) / 72
         dy = (self._labelsep_inches + fontsize) / 2
         trans = offset_copy(ax.transAxes, fig, 0, dy)  # type: ignore
         self.set_transform(trans)
@@ -441,8 +455,8 @@ class Compass(PathCollection):
         angle: float | None = None,
         size: float = 20,
         style: Literal["arrow", "star", "circle"] = "arrow",
-        pc_kwargs: dict[str, Any] | None = None,
-        text_kwargs: dict[str, Any] | None = None,
+        pc_kwargs: CompassPcKwargs | None = None,
+        text_kwargs: CompassTextKwargs | None = None,
     ) -> None:
         self.x = x
         self.y = y
@@ -479,6 +493,7 @@ class Compass(PathCollection):
                 )
 
         pc_kwargs = normalize_kwargs(pc_kwargs, PathCollection)  # type: ignore
+        pc_kwargs = cast(CompassPcKwargs, pc_kwargs)
         pc_kwargs.setdefault("linewidth", 1)
         pc_kwargs.setdefault("edgecolor", "k")
         pc_kwargs.setdefault("facecolor", colors)
@@ -487,9 +502,10 @@ class Compass(PathCollection):
         super().__init__(paths, transform=None, **pc_kwargs)
 
         # 文字在箭头上方
-        pad = head / 2.5
         text_kwargs = normalize_kwargs(text_kwargs, Text)  # type: ignore
+        text_kwargs = cast(CompassTextKwargs, text_kwargs)
         text_kwargs.setdefault("fontsize", size / 1.5)
+        pad = head / 2.5
         self.text = Text(
             x=0,
             y=axis + pad,
@@ -498,7 +514,7 @@ class Compass(PathCollection):
             va="center",
             rotation=0,
             transform=None,
-            **text_kwargs,
+            **text_kwargs,  # type: ignore
         )
 
     @staticmethod
@@ -646,13 +662,14 @@ class ScaleBar(Axes):
 class Frame(Artist):
     """GMT风格的边框"""
 
-    def __init__(self, width: float = 5, **kwargs: Any) -> None:
+    def __init__(self, width: float = 5, **kwargs: Unpack[FrameKwargs]) -> None:
         self.width = width
         self._width = width / 72
         super().__init__()
         self.set_zorder(2.5)
 
-        kwargs = normalize_kwargs(kwargs, PathCollection)
+        kwargs = normalize_kwargs(kwargs)  # type: ignore
+        kwargs = cast(FrameKwargs, kwargs)
         kwargs.setdefault("linewidth", 1)
         kwargs.setdefault("edgecolor", "k")
         kwargs.setdefault("facecolor", ["k", "w"])