+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.0 (ongoing)
 
-- 新增 `typing_extensions` 依赖，为 `python==3.10` 提供类型提示。
+- 新增 `typing_extensions>=4.13.0` 依赖，为 `python==3.10` 提供类型提示。
+- 通过 `TypedDict` 和 `Unpack` 帮传给 matplotlib 的 `**kwargs` 参数提供类型提示。
 - 彻底弃用 `option` 模块。
 
 ## 0.7.6 (2025-10-10)